+++ src/comuSasoi.c
@@ -535,7 +535,7 @@ u8 FUN_003c2c50(u32 param_1)
 
   lVar4 = FUN_00172c50(&sStack_2);
 
-  if ((lVar4 == 0) || (param_1 != (long)sStack_2)) {
+  if ((lVar4 == 0) || (param_1 != sStack_2)) {
 
     uVar5 = (u16)param_1;
